add hone work 2, add webpack congfig in project, fix eslint bugs, crushed on components
