+++ .eslintrc.yml
@@ -1,3 +1,5 @@
+parserOptions:
+  sourceType: 'module'
 env:
   es6: true
   browser: true
