+++ public/index.html
@@ -39,6 +39,6 @@
         <h1 class="control__title">TASKMANAGER</h1>
       </section>
     </main>
-    <script src="../src/main.js"></script>
+    <script src="./bungle.js"></script>
   </body>
 </html>
