+++ src/components/board-container.js
@@ -1,0 +1,5 @@
+// ф-ция возвращает главный контейнер плриложения ".board container"
+export const markupMainBoardContainer = () => {
+  return `<section class="board container">
+  </section>`;
+};
