+++ src/components/board-tasks.js
@@ -1,0 +1,5 @@
+// ф-ция возвращает контейнер для крточек ".board__tasks"
+ export const markupBoardTasks = () => {
+  return `<div class="board__tasks">
+  </div>`;
+};
