+++ src/components/load-more-btn.js
@@ -1,0 +1,4 @@
+// ф-ция создвния кнопки "load moreconst markupLoadMore = () => {
+export const markupLoadMore = () => {
+  return `<button class="load-more" type="button">load more</button>`;
+};
